Adding basic tests, cleaning up servers and package json
This commit adds initial tests just to ensure that the server responds properly
to a test. More tests still need to be written

This also cleans up the package json and server files to remove some things that
aren't actually being used at this point. Also updates the gitignore to ignore
the node_modules folder