Betting handling of change events for code mirror
This change updates how change events are emitted and rendered from/onto the
client. Previously it was happening on document key up, which was limiting how
users could do things like command-a to highlight all.

Experimented with concepts such as updating on a set interval, but found that to
be ineffective and not solving the actual problem