Adding the ability to create and access code sessions
This adds the ability to create and access code sessions. Adds server endpoints
and tweaks the front end so that users are directed to their session upon creation.

Additionally, it tweaks the rendering of a codemirror so that the text is loaded
automatically and the code is saved every 5 seconds.

There's also some minor styling adjustments in here